optimize tsx loading
