+++ packages/wdio-config/package.json
@@ -39,7 +39,6 @@
     "@wdio/logger": "workspace:*",
     "@wdio/types": "workspace:*",
     "@wdio/utils": "workspace:*",
-    "decamelize": "^6.0.0",
     "deepmerge-ts": "^7.0.3",
     "glob": "^10.2.2",
     "import-meta-resolve": "^4.0.0"
